issue-561: don't cast function pointer to void *
Which gcc-3.4 (as shipped in rhel 4) doesn't like.

Cast to void * was originally added to avoid issue on OSX which
doesn't have sighandler_t.

In that place we only need to know if it's null or not. So casting to
intptr_t looks like simplest possible way to achieve that.