issue-587: fix typos in unit test scripts
As proposed by Hannes Weisbach.

The argument will be garbled because of a misplaced brace, for example
(heap-checker_unittest.sh):

HEAP_CHECKER="${1:-$BINDIR}/heap-checker_unittest"
which should be:
HEAP_CHECKER="${1:-$BINDIR/heap-checker_unittest}"

This unit test is used to check the binaries heap-checker_unittest and
heap-checker_debug_unittest.  With the typo, the executable
heap-checker_debug_unittest is never actually run.