issue-588: Fix profiler_unittest.cc fork()
As suggested by Hannes Weisbach.

Call heap-profiler_unittest with the arguments 1 -2 (one iteration, 2
fork()ed children).

Instead of running the test, the program crashes with a std::bad_alloc
exception.  This is caused by unconditionally passing the
number-of-threads-argument (0 or positive for threads, negative for
fork()s) in RunManyThreads(), thus allocating an array of pthread_t of
size -2.  Depending on the sign of the thread number argument either
RunManyThreads or fork() should be called.