PowerPC: stacktrace function refactor and fixes
This patch fixes the stacktrace creating when the function is
interrupted by a signal. For Linux, the vDSO signal trampoline symbol is
compared against LR from stack backchain and handled different in that
case (since the signal trampoline layout a different stack frame).

Because of this extensive change the PowerPC stacktrace code has now
been refactored to split in Linux and Darwin specific codes.