Fix parsing /proc/pid/maps dump in CPU profile data file
When trying to use pprof on my machine, the symbols of my program were
not being recognized.

It turned out that pprof, when calculating the offset of the text list
of mapped objects (the last section of the CPU profile data file), was
assuming that the slot size was always 4 bytes, even on 64-bit machines.

This led to ParseLibraries() reading a lot of garbage data at the
beginning of the map, and consequently the regex was failing to match on
the first line of the real (non-garbage) map.