pprof: eliminate duplicate top frames if dropping signal frames
In cpu profiles that had parts of signal handler we could have
situation like that:

* PC
* signal handler frame
* PC

Specifically when capturing stacktraces via libunwind.

For such stacktraces pprof used to draw self-cycle in functions
confusing everybody. Given that me might have a number of such
profiles in the wild it makes sense to treat that duplicate PC issue.