rename sys allocator's sys_alloc symbol to tcmalloc_sys_alloc
Since we're not building with hidden visibility, symbols of other DSOs
can interpose our symbols. And sys_alloc is just too generic name.

And in fact erlang runtime has sys_alloc function. Which means we're
trying to write to it's first bytes as part of initializing system
allocator and crash.

This should fix issue #843.