Support viewing profiles in callgrind with templates
The new regex looks for the -> separator instead of a space. This is
important when using --no_strip_temp, because (at least on my system)
the demangler likes putting spaces between template arguments and
in things like `foo<bar<baz> >`. I say "on my system" because that seems
like the kind of thing that might change once C++11 becomes widely
adopted.