Port GWP-ASan from internal TCMalloc.
GWP-ASan is a low-overhead sampling-based utility for finding
heap-use-after-frees and heap-buffer-overflows.

This implementation supports up to 64 allocations at a time with sizes
in the range [1, kPageSize]. The sampling rate is controlled by the
environment variable TCMALLOC_GUARDED_SAMPLE_PARAMETER, which defaults
to never doing a guarded allocation.

When a memory error (segfault) is detected, stack traces are printed for
point of allocation, point of deallocation (if applicable), and point of
error.  This implementation crashes on segfault.