doubly-linked free-lists for thread caches
Added the ability for free lists to be built out of doubly-linked
lists in tcalloc. TCMALLOC_USE_DOUBLYLINKED_FREELIST flag must be set
in order for doubly-linked lists to be used.
By default flag is only set in debug allocation builds.

These changes are based on the following Chromium commits:
- https://codereview.chromium.org/7671034
- http://codereview.chromium.org/8000001
- http://codereview.chromium.org/10262029
- https://chromiumcodereview.appspot.com/10944024
- https://chromiumcodereview.appspot.com/10957067
- https://chromiumcodereview.appspot.com/11362046
- https://chromiumcodereview.appspot.com/12619004