syscall: Mark x8 as clobbered
Mark arm64 register x8 as clobbered by syscall body inline assembly as
it is being used to store the syscall number. Otherwise the compiler
may try to use it for some other purpose.

This fix is derived from a resolution to clang Bugzilla report
https://bugs.llvm.org/show_bug.cgi?id=48798.  See this report for a
minimal reproducer derived from the code fixed here as well as the
resolution.

This should fix SEGFAULTs as reported in
https://github.com/envoyproxy/envoy/issues/14756.

Fixes: #1241
Signed-off-by: Jarno Rajahalme <jarno@covalent.io>