stacktrace_unittest: test backtracing from signals more reliably
Somehow I originally choose to segfault to test this, and it caused us
to deal with 'skipping over' null pointer access. Which ended up not
so easy to do semi-portably. But easier way is to do same
itimer/sigprof thing that we do in actual profiler.

We're still somewhat at the mercy of compiler placing code "normally",
but this implementation seems more robust anyways.