Fixed an error in ppref
When searching for symbols in the analyzed program, if the program is analyzed
on another machine, the paths to the program executable file specified in the
pprof call argument and the path specified in the analysis results file may
differ. Therefore, the condition of comparing these paths does not work. It is
necessary to compare not the paths, but the names of executable files.