introduce 2 new env vars to config ThreadCache initial state
new variables:
TCMALLOC_BATCH_SIZE_FROM_START:
Set the thread cache max length size to num_objects_to_move(cl) if defined

The default behavior is to set the initial max_length to 1 and increment it by 1 each time
objects are fetched. Once the max_lenght reaches num_objects_to_move(cl), the max_length is now
incremented by num_objects_to_move(cl) as it has been demonstrated that this particular size is in high
demand.

This is done that way to minimize memory wasting. Bigger blocks are allocated only once the need for
this size class has been confirmed.

If the possibility of wasting some memory is acceptable, enabling this option to eliminate the ramp-up period will:
- reduce number of accesses to the central heap
- Preserve the data locality of the objects in opposition to have the address space distributed evenly
  across all the caches if a particular size is requested by all threads.

This second point is not only valid when the client code access the allocated memory but also when the memory
is deallocated when it is put back into a free list.

TCMALLOC_INITIAL_THREAD_CACHE_BYTES:
Set the initial cache max_size_ value.

The default behavior is to initialize the value to kStealAmount (64Kb) and increment it by kStealAmount
each time a Scavenge() is triggered when a cache size reach the max_size value. This approach is the best to
efficiently distribute the heap across all the caches but this imply several accesses to the central heap.

If you know the number of created threads by a process this env variable allows you to manually distribute the heap
across all the threads and eliminates several central heap accesses needed to reach their stable size.

Signed-off-by: Olivier Langlois <olivier@trillion01.com>