[low_level_alloc] allocate Arena with custom allocator when provided
This refactors API so that meta_data_arena isn't needed or used
anymore. We always used nullptr for this argument anyways.

We also change code so that PagesAllocator API is able to choose
rounding up of allocation requests. It now returns pointer to
allocated memory and it's rounded up size.

And most important part of this change is that Arena instance itself
is now allocated with provided allocator (when provided). This reduces
locking depencies.

We make sure to cover new logic by test.