[tcmalloc_unittest] don't disable threaded test for fork torturing
We now detect "SIGTRAP is blocked" case and handle it properly. This
enables us to run pthread_create. So there is no more need to disable
ManyThreads test.

Note, we're not propagating single-stepping mode into child threads,
so we're still only fork-testing main thread, but that should be
plenty good for now. I.e. merely not crashing on pthread_create
anymore when fork torturing is enabled.