simplify away writing of tcmalloc.h from tcmalloc.h.in
We had some degree of mess caused by need to duplicate tcmalloc.h for
windows and non-windows and caused by duplication of rewriting logic
between cmake stuff and autotools stuff.

We only need it to handle availability of mallinfo stuff, which we're
able to mostly detect by looking at various libc defines.

And we need it for the version defines. Which we were already
partially hardcoding for each version bump anyways. So we simply add a
test that version hardcoded in tcmalloc.h header is the same as in our
autoconf bits.