introduce "generic" config.h header
This contains roughly same defines as autotools-produced config.h, but
it checks headers with GCC-specific tests and it checks some features
by inspecting known platform defines (i.e. like __GLIBC__). So by
definition it supports fewer platforms, but it is "good enough" for
commonly used ones. It is to be used in our bazel support (and maybe
more later).