tcmalloc_unittest: unbreak sub-process spawning on windows
Turns out spawnve (which is part of ucrt) has a bug. They have some
sort of seemingly undocumented thing where current working
directory (or maybe history of working directories) is preserved in
special format inside environment block. And spawnve tries to copy it
and fails to account for possibility of not having this
stuff.

https://github.com/huangqinjin/ucrt/blob/d6e817a4cc90f6f1fe54f8a0aa4af4fff0bb647d/exec/cenvarg.cpp#L136

Apparently when bazel runs tests it clears environment so there is
nothing to be found and so their naive loop does past-malloced-block
access and crashes.

We avoid the it by using somewhat more direct CreateProcessA API.