make GetenvBeforeMain simply use normal getenv
It is actually okay to call very early (with notable exception of
windows, so we keep part of the madness).

With this we can also amputate totally unneccessary __environ
thingy. Hard to say if it was ever necessary, but as far as my code
archeology goes, using regular getenv (or environ) should have been
safe since long ago.