tcmalloc_unittest: un-flake the large allocs release test
FreeBSD mmap sys-allocator is behaving "randomly" enough that old
logic wasn't triggering the right sequences of addresses.

So I had to rework the test a bit to actually ensure that we're
allocating used/free addressess one exactly after the other.

I also previously added stupid duplicate release rate reset, which was
already there, but I was somehow blind enough not to spot it just 1
line below! So this is removed.

And this entire test is now separate function with hopefully helpful
comments about what we're trying to achieve.

Updates github #1604 and #1574